// frontend/apis/model/userSignUpInput.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Budget Calendar Service
 * Budget Calendar API
 * OpenAPI spec version: 1.0
 */

/**
 * Sign Up Input
 */
export interface UserSignUpInput {
  /** ユーザー名 */
  name: string;
  /** メールアドレス */
  email: string;
  /** パスワード */
  password: string;
}